requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
python-telegram-bot==21.0.1
jinja2
requests
python-multipart
google-cloud-firestore
